+++ src/main/java/com/java/cruisecompany/model/entity/User.java
@@ -3,7 +3,8 @@ package com.java.cruisecompany.model.entity;
 import com.java.cruisecompany.model.entity.enums.Role;
 import lombok.Builder;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
+
+import java.math.BigDecimal;
 
 @Builder
 @Data
@@ -15,5 +16,5 @@ public class User {
     private String firstName;
     private String lastName;
     private Role role;
-    private double balance;
+    private BigDecimal balance;
 }
